Creates the input component
Connects that component to the redux store so that the App component does not need to be connected to redux
