+++ src/App.js
@@ -1,32 +1,13 @@
-import "./App.scss";
-
 import React, { Component } from "react";
+import "./App.scss";
+import Input from "./component/input/Input";
 
-import { setInput } from "./redux/input/input.actions";
-import { connect } from "react-redux";
-
-const mapStateToProps = (state) => {
-    return {
-        input: state.input.content,
-    };
-};
-
-const mapDispatchToProps = (dispatch) => {
-    return {
-        set: (content) => dispatch(setInput(content)),
-    };
-};
-
-export class App extends Component {
-    onChange = (e) => {
-        this.props.set(e.target.value);
-    };
-
+class App extends Component {
     render = () => (
         <div className="App">
-            <input onChange={this.onChange}></input>
+            <Input />
         </div>
     );
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(App);
+export default App;
