+++ src/component/input/Input.jsx
@@ -1,0 +1,29 @@
+import React, { Component } from "react";
+import { connect } from "react-redux";
+import { setInput } from "../../redux/input/input.actions";
+
+const mapStateToProps = (state) => {
+    return {
+        input: state.input.content,
+    };
+};
+
+const mapDispatchToProps = (dispatch) => {
+    return {
+        set: (content) => dispatch(setInput(content)),
+    };
+};
+
+class Input extends Component {
+    onChange = (e) => {
+        this.props.set(e.target.value);
+    };
+
+    render = () => (
+        <div className="input-parent-container">
+            <input onChange={this.onChange}></input>
+        </div>
+    );
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(Input);
